src/requirements-dev.txt
black==25.1.0
celery-stubs==0.1.3
django-stubs==5.2.0
django-stubs-ext==5.2.0
djangorestframework-stubs==3.16.0
flake8==7.2.0
Flake8-pyproject==1.2.3
freezegun==1.5.1
isort==6.0.1
mimesis==18.0.0
mypy==1.15.0
mypy-extensions==1.1.0
pytest==8.3.5
pytest-blockage==0.2.4
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-lazy-fixtures==1.1.2
pytest-mock==3.14.0
types-pytz==2025.2.0.20250326
types-requests==2.32.0.20250328
pytest-xdist==3.6.1
psutil==7.0.0
execnet==2.1.1